python-server/pyproject.toml
[project]
name = "python-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.7.6",
    "auth>=0.5.3",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.11.0.post1",
    "fastapi>=0.116.1",
    "fastmcp<2.10.4",
    "google-api-python-client>=2.177.0",
    "google-auth-oauthlib>=1.2.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-google-genai>=2.1.8",
    "langchain-groq>=0.3.6",
    "mcp[cli]>=1.9.1",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pypdf>=5.8.0",
    "pytesseract>=0.3.13",
    "rank-bm25>=0.2.2",
    "tzlocal>=5.3.1",
    "uvicorn>=0.35.0",
]
